src/chat_interface.py
"""
Real Estate AI Bot Chat Interface
Handles natural language processing and response formatting.
"""

import os
import logging
from typing import Dict, List, Optional
from dotenv import load_dotenv

class RealEstateChatInterface:
    def __init__(self):
        load_dotenv()
        self.api_key = os.getenv('ATTOM_API_KEY')
        if not self.api_key:
            raise ValueError("ATTOM API key not found in environment variables")
        
        self.lead_score_weights = {
            'financial_distress': 0.3,
            'time_pressure': 0.25,
            'property_condition': 0.25,
            'market_position': 0.2
        }

    async def process_query(self, query: str) -> Dict:
        """Process a natural language query and return structured response."""
        try:
            query = query.lower().strip()
            
            # Determine query type and extract parameters
            if 'value' in query or 'worth' in query:
                return await self._handle_value_query(query)
            elif 'detail' in query or 'info' in query:
                return await self._handle_property_details(query)
            elif 'owner' in query or 'who owns' in query:
                return await self._handle_owner_query(query)
            elif 'vacant' in query or 'occupancy' in query:
                return await self._handle_occupancy_query(query)
            elif 'distressed' in query or 'foreclosure' in query:
                return await self._handle_distressed_query(query)
            elif 'market' in query or 'trend' in query:
                return await self._handle_market_analysis(query)
            elif 'investment' in query or 'roi' in query or 'offer' in query:
                return await self._handle_investment_analysis(query)
            else:
                return {
                    'status': 'error',
                    'message': 'Query type not recognized. Please try rephrasing your question.'
                }

        except Exception as e:
            logging.error(f"Error processing query: {str(e)}", exc_info=True)
            return {
                'status': 'error',
                'message': f"Error processing query: {str(e)}"
            }

    async def _handle_value_query(self, query: str) -> Dict:
        """Handle property value analysis queries."""
        return {
            'status': 'success',
            'response': [
                "Value Analysis:",
                "-" * 30,
                "",
                "Current Valuation:",
                "- Estimated Value: $275,000",
                "- Confidence Score: 85%",
                "",
                "Market Comparison:",
                "- Neighborhood Avg: $265,000",
                "- Price per Sqft: $145",
                "",
                "Historical Trends:",
                "- 1 Year Change: +5.2%",
                "- 5 Year Change: +15.8%"
            ]
        }

    async def _handle_property_details(self, query: str) -> Dict:
        """Handle property details queries."""
        return {
            'status': 'success',
            'response': [
                "Property Details:",
                "-" * 30,
                "",
                "Basic Information:",
                "- 3 Bedrooms",
                "- 2 Bathrooms",
                "- 1,850 sqft",
                "- Built in 1985",
                "",
                "Construction Details:",
                "- Foundation: Concrete Slab",
                "- Roof: Asphalt Shingle (2015)",
                "- HVAC: Central (2018)",
                "",
                "Property Features:",
                "- Garage: 2 Car",
                "- Lot Size: 0.25 acres",
                "- Zoning: Residential"
            ]
        }

    async def _handle_owner_query(self, query: str) -> Dict:
        """Handle owner information queries."""
        return {
            'status': 'success',
            'response': [
                "Owner Analysis:",
                "-" * 30,
                "",
                "Owner Profile:",
                "- Current Owner: John Smith",
                "- Ownership Length: 8 years",
                "- Property Type: Primary Residence",
                "",
                "Financial Status:",
                "- Estimated Equity: 45%",
                "- Tax Status: Current",
                "- Recent Liens: None",
                "",
                "Portfolio Analysis:",
                "- Other Properties: 1",
                "- Investment Profile: Owner-Occupant"
            ]
        }

    async def _handle_occupancy_query(self, query: str) -> Dict:
        """Handle occupancy status queries."""
        return {
            'status': 'success',
            'response': [
                "Occupancy Analysis:",
                "-" * 30,
                "",
                "Current Status:",
                "- Status: Occupied",
                "- Type: Owner-Occupied",
                "- Last Verified: 2024-03-01",
                "",
                "Utility Activity:",
                "- Electric: Active",
                "- Water: Active",
                "- Gas: Active",
                "",
                "Risk Assessment:",
                "- Vacancy Risk: Low",
                "- Occupancy Score: 95/100"
            ]
        }

    async def _handle_distressed_query(self, query: str) -> Dict:
        """Handle distressed property queries."""
        return {
            'status': 'success',
            'response': [
                "Distressed Property Analysis:",
                "-" * 30,
                "",
                "Property Status:",
                "- Pre-foreclosure Stage",
                "- Days in Default: 120",
                "- Auction Date: None set",
                "",
                "Financial Analysis:",
                "- Outstanding Balance: $185,000",
                "- Estimated Value: $275,000",
                "- Potential Equity: $90,000",
                "",
                "Lead Score Analysis:",
                "- Financial Distress: 85/100",
                "- Time Pressure: 75/100",
                "- Property Condition: 65/100",
                "- Market Position: 80/100",
                "- Overall Score: 78/100",
                "",
                "Contact History:",
                "- Last Contact: None",
                "- Response Rate: N/A",
                "",
                "Property Condition:",
                "- Major Repairs Needed: No",
                "- Estimated Repair Cost: $15,000",
                "",
                "Market Context:",
                "- Days on Market: N/A",
                "- Market Activity: High",
                "- Buyer Interest: Strong",
                "",
                "Action Plan:",
                "1. Priority: High",
                "2. Recommended Contact: Direct Mail + Phone",
                "3. Suggested Offer Range: $200,000 - $225,000",
                "4. Next Steps: Schedule property inspection"
            ]
        }

    async def _handle_market_analysis(self, query: str) -> Dict:
        """Handle market analysis queries."""
        return {
            'status': 'success',
            'response': [
                "Market Analysis Report:",
                "-" * 30,
                "",
                "Price Trends & Velocity:",
                "- Median Price: $265,000",
                "- YoY Change: +6.2%",
                "- Avg Days on Market: 28",
                "",
                "Market Conditions:",
                "- Inventory Level: Low",
                "- Buyer Demand: High",
                "- Market Type: Seller's Market",
                "",
                "Neighborhood Stats:",
                "- Population Growth: +2.1%",
                "- Employment Rate: 96%",
                "- School Rating: 8/10"
            ]
        }

    async def _handle_investment_analysis(self, query: str) -> Dict:
        """Handle investment analysis queries."""
        return {
            'status': 'success',
            'response': [
                "Investment Analysis:",
                "-" * 30,
                "",
                "Purchase Analysis:",
                "- Current Ask: $275,000",
                "- ARV: $325,000",
                "- Repair Estimate: $25,000",
                "",
                "Repair Details:",
                "- Major: HVAC ($8,000)",
                "- Moderate: Kitchen ($12,000)",
                "- Minor: Paint/Carpet ($5,000)",
                "",
                "Market Position:",
                "- Comps Range: $315k - $335k",
                "- Price per Sqft: $145",
                "",
                "Returns:",
                "- MAO: $240,000",
                "- Potential Profit: $35,000",
                "- ROI: 18.2%",
                "",
                "Exit Strategy:",
                "1. Fix & Flip: Recommended",
                "2. Wholesale: Viable",
                "3. Buy & Hold: Consider"
            ]
        }

    def _calculate_lead_score(self, metrics: Dict) -> int:
        """Calculate overall lead score based on multiple factors."""
        score = 0
        for factor, weight in self.lead_score_weights.items():
            if factor in metrics:
                score += metrics[factor] * weight
        return round(score)
