requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
